+++ lib/request_permission.dart
@@ -1,6 +1,4 @@
 library request_permission;
 
+export 'src/android_permissions.dart';
 export 'src/request_permission.dart';
-
-/// Has to be finished
-// export 'src/android_permissions.dart';